+++ stdlib/Dates/src/accessors.jl
@@ -81,7 +81,7 @@ julia> isoyear(Date(2021, 12, 31))
 !!! compat "Julia 1.13"
     This function requires Julia 1.13 or later.
 """
-function isoyear(dt::DateTime)
+function isoyear(dt::Union{DateTime,Timestamp})
     thisyear = Year(dt)
     thismonth = Month(dt)
     weeknumber = week(dt)
@@ -101,7 +101,8 @@ isoyear(dt::Date) = isoyear(DateTime(dt))
 Return the ISO week date that corresponds to `dt` (see
 https://en.wikipedia.org/wiki/ISO_week_date).
 
-The return type is a tuple of `Year`, `Week` and `Int64` (from 1 to 7).
+The return type is a tuple of three `Int64` values: the ISO year, the ISO week
+number, and the ISO weekday (from 1 through 7).
 
 # Examples
 ```jldoctest
@@ -114,7 +115,7 @@ julia> isoweekdate(Date(2023, 01, 01))
 !!! compat "Julia 1.13"
     This function requires Julia 1.13 or later.
 """
-isoweekdate(dt::DateTime) = (isoyear(dt).value, week(dt), dayofweek(dt))
+isoweekdate(dt::Union{DateTime,Timestamp}) = (isoyear(dt).value, week(dt), dayofweek(dt))
 isoweekdate(dt::Date) = isoweekdate(DateTime(dt))
 
 function quarter(days)
@@ -125,9 +126,15 @@ end
 
 # Accessor functions
 value(dt::TimeType) = dt.instant.periods.value
+value(dt::Timestamp) = value(dt.instant.periods)
 value(t::Time) = t.instant.value
 days(dt::Date) = value(dt)
 days(dt::DateTime) = fld(value(dt), 86400000)
+days(dt::Timestamp{P}) where {P} = fld(value(dt), timestamp_ticks_per_day(P)) + UNIXEPOCHDAYS
+# Nanoseconds since midnight
+nsofday(dt::Date) = Int64(0)
+nsofday(dt::DateTime) = 1000000 * mod(value(dt), 86400000)
+nsofday(dt::Timestamp{P}) where {P} = mod(value(dt), timestamp_ticks_per_day(P)) * timestamp_scale(P)
 year(dt::TimeType) = year(days(dt))
 quarter(dt::TimeType) = quarter(days(dt))
 month(dt::TimeType) = month(days(dt))
@@ -143,6 +150,17 @@ second(t::Time) = mod(fld(value(t), 1000000000), Int64(60))
 millisecond(t::Time) = mod(fld(value(t), Int64(1000000)), Int64(1000))
 microsecond(t::Time) = mod(fld(value(t), Int64(1000)), Int64(1000))
 nanosecond(t::Time) = mod(value(t), Int64(1000))
+# A time field of `dt` that is `unit` nanoseconds long, counted modulo `modulus`
+# (24 for hours). Fields finer than P are zero.
+@inline timestamp_part(dt::Timestamp{P}, unit, modulus) where {P} =
+    unit < timestamp_scale(P) ? Int64(0) :
+    mod(fld(value(dt), unit ÷ timestamp_scale(P)), modulus)
+hour(dt::Timestamp) = timestamp_part(dt, 3600000000000, Int64(24))
+minute(dt::Timestamp) = timestamp_part(dt, 60000000000, Int64(60))
+second(dt::Timestamp) = timestamp_part(dt, Int64(1000000000), Int64(60))
+millisecond(dt::Timestamp) = timestamp_part(dt, Int64(1000000), Int64(1000))
+microsecond(dt::Timestamp) = timestamp_part(dt, Int64(1000), Int64(1000))
+nanosecond(dt::Timestamp) = timestamp_part(dt, Int64(1), Int64(1000))
 
 dayofmonth(dt::TimeType) = day(dt)
 
@@ -157,7 +175,7 @@ for func in (:year, :month, :quarter)
         @doc """
             $($name)(dt::TimeType)::Int64
 
-        The $($name) of a `Date` or `DateTime` as an [`Int64`](@ref).
+        The $($name) of a `Date`, `DateTime`, or `Timestamp` as an [`Int64`](@ref).
         """ $func(dt::TimeType)
     end
 end
@@ -165,11 +183,11 @@ end
 """
     week(dt::TimeType)::Int64
 
-Return the [ISO week date](https://en.wikipedia.org/wiki/ISO_week_date) of a `Date` or
-`DateTime` as an [`Int64`](@ref). Note that the first week of a year is the week that
-contains the first Thursday of the year, which can result in dates prior to January 4th
-being in the last week of the previous year. For example, `week(Date(2005, 1, 1))` is the 53rd
-week of 2004.
+Return the [ISO week number](https://en.wikipedia.org/wiki/ISO_week_date) of a
+`Date`, `DateTime`, or `Timestamp` as an [`Int64`](@ref). Note that the first
+week of a year is the week that contains the first Thursday of the year, which
+can result in dates prior to January 4th being in the last week of the previous
+year. For example, `week(Date(2005, 1, 1))` is the 53rd week of 2004.
 
 # Examples
 ```jldoctest
@@ -191,7 +209,7 @@ for func in (:day, :dayofmonth)
         @doc """
             $($name)(dt::TimeType)::Int64
 
-        The day of month of a `Date` or `DateTime` as an [`Int64`](@ref).
+        The day of month of a `Date`, `DateTime`, or `Timestamp` as an [`Int64`](@ref).
         """ $func(dt::TimeType)
     end
 end
@@ -221,8 +239,8 @@ for parts in (["year", "month"], ["month", "day"], ["year", "month", "day"])
         @doc """
             $($name)(dt::TimeType) -> ($(join(repeated(Int64, length($parts)), ", ")))
 
-        Simultaneously return the $(join($parts, ", ", " and ")) parts of a `Date` or
-        `DateTime`.
+        Simultaneously return the $(join($parts, ", ", " and ")) parts of a `Date`,
+        `DateTime`, or `Timestamp`.
         """ $func(dt::TimeType)
     end
 end
@@ -231,9 +249,9 @@ for func in (:hour, :minute, :second, :millisecond, :microsecond, :nanosecond)
     name = string(func)
     @eval begin
         @doc """
-            $($name)(t::Time)::Int64
+            $($name)(t::Union{Time,Timestamp})::Int64
 
-        The $($name) of a `Time` as an [`Int64`](@ref).
+        The $($name) of a `Time` or `Timestamp` as an [`Int64`](@ref).
         """ $func(t::Time)
     end
 end